requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
supabase==1.2.0
openai==1.12.0
pydantic==2.6.1
pydantic-settings==2.1.0
tiktoken==0.5.2
email-validator
pandas==2.2.0
python-multipart==0.0.6